+++ functions/middlewares.js
@@ -65,7 +65,7 @@ module.exports.chargeIfNeeded = (req, res, next) => {
         admin.database().ref('/prices/' + req.mnData.nodeType)
             .once('value', snapshot => {
                 const amount = snapshot.val()
-                const dailyAmount = parseFloat(amount) / 30 // dividing by 30 so we can get the daily rate
+                const dailyAmount = parseInt(amount) / 30 // dividing by 30 so we can get the daily rate
 
                 updateBalance(req.orderData.userId, dailyAmount * -1,
                     async (err) => {
